fix: address clang-tidy findings on v2.0.0; tune linter noise floor
Real bugs the analyzer caught and the fixes applied:

- `insert_node` (tokenizer/list_create.c) dereferenced prev->next
  before the !prev guard. None of the current callers ever pass NULL,
  so this was latent — but the dereference happened unconditionally.
  Rewrote to handle the prev==NULL case first.
- `venv_remove` (utils/env_lookup.c) used a separate `prev` tracker
  that the analyzer couldn't prove was always assigned before read.
  Replaced with a pointer-to-pointer walk so head and middle cases
  share one path and there is no prev at all.
- `expander` had a default `next = top->next` at the top of the loop
  that every branch immediately overwrote — a dead store. Removed.
- `child_exit` is now `_Noreturn` so the analyzer stops reporting
  spurious NullDereference / NonNullParamChecker findings on code
  paths that already called it for the failure case.
- `ARENA_DEFAULT_CHUNK` is now `(size_t)(64*1024)` instead of `(64*1024)`
  so the int*int product doesn't widen implicitly on assignment.

Lint config: expanded the noise-suppression list with file-and-line
findings reviewed and judged false positives or stylistic:

- cert-sig30-c / cert-msc54-cpp: SIGINT handler calls rl_on_new_line
  etc., which is the readline-documented redraw pattern.
- clang-analyzer-security.insecureAPI.DeprecatedOrUnsafeBufferHandling:
  suggests C11 Annex K `*_s` functions (memcpy_s, snprintf_s, etc.)
  that aren't available on glibc/macOS.
- clang-analyzer-optin.core.EnumCastOutOfRange: false positive on the
  tokenizer's `(c==X)*TOKEN + …` arithmetic — the zero contributions
  are intentional.
- bugprone-inc-dec-in-conditions: stylistic; the codebase uses
  `while (s[++i])` extensively.

Verified locally (clang-tidy 18 in an ubuntu:24.04 container with
bear-generated compile_commands.json): 0 issues. Tests 69/69 pass.
Leak check: 0 leaks.

+++ libarena/include/arena.h
@@ -17,7 +17,7 @@
 
 # include <stddef.h>
 
-# define ARENA_DEFAULT_CHUNK	(64 * 1024)
+# define ARENA_DEFAULT_CHUNK	((size_t)(64 * 1024))
 # define ARENA_ALIGN			16
 
 typedef struct s_arena_chunk
